Branch-based tagging with weekly builds.
Enables smart docker tagging based on git branches and git tags. Builds are ran automatically on 'push' events. The docker tags will be based on the branch that was pushed. For example, if 'master' was pushed, then the standard tags will get created such as 'latest' and 'x86_64'. If there is a tag on HEAD, then it will also generate the version-based tags, like 'v1.1-x86_64' and 'v1.1'.  The builds are also ran weekly for the master branch, pulling in any security updates from the base layers.

Other changes:
* Improved README.md with more details about the project and links to docker hub
* Added docker-compose.yaml for easier image building
* Added failure flags to curl to ensure it exits on error

Signed-off-by: Daniel <daniel@developerdan.com>